adopting compressed agent concept
This also references #12.

Further testing and comparison to what browscap-php is doing, I noticed
that browscap-php strips digits out of the pattern and replaces them
with “(\d)” and then only stores the single pattern for all the
different version combinations (they also store the version
combinations that can be matched, and then use those later to verify
that things like “iOS 89.5454” don’t match one that is meant for
something like “iOS 8.1”).

Duplicating this compression idea in crossjoin, which results in a LOT
less patterns stored in the cache files.  I’m not storing the version
combinations, but instead I’m continuing the loop in getBrowser if the
pattern with the filled in digits doesn’t match any of the ini data
stored in the iniparts cache files, which appears to work as expected.
Storing the version combinations seems unnecessary.

With this change, I’m able to make crossjoin outperform browscap ~10x
in the different suites I used.

These two suites are included with the user-agent-parser-benchmarks
project.

>Running Benchmark using agents in ua-list-100-sample01.txt
>Benchmarking browscap-php ...	14.775481secs 	35.25MB
>Benchmarking crossjoin-browscap ...	0.8129889secs 	1.25MB

>Running Benchmark using agents in ua-list-all.txt
>Benchmarking browscap-php ...	72.423168secs 	35.25MB
>Benchmarking crossjoin-browscap ...	6.3292379secs 	2.5MB

These two files are collections of user agents from our system:

>Running Benchmark using agents in top-1000-last-7days.txt
>Benchmarking browscap-php ...	183.35667secs 	35.75MB
>Benchmarking crossjoin-browscap ...	24.990783secs 	4.5MB

>Running Benchmark using agents in top-200-last-24.txt
>Benchmarking browscap-php ...	51.871809secs 	35.25MB
>Benchmarking crossjoin-browscap ...	5.9958388secs 	1.75MB

This text file just has the worst performing user agent from the file
above (top 200) that I mentioned in #12:

>Running Benchmark using agents in problem-ua.txt
>Benchmarking browscap-php ...	0.1463539secs 	35MB
>Benchmarking crossjoin-browscap ...	0.0426089secs 	0.5MB

Here’s a profile of that last run, compare to the profile in #12:

https://blackfire.io/profiles/4ba82060-fc69-4c77-8fc0-31635b90070f/graph

I used the benchmark utility’s tool for comparing the diff between the
parsing libraries, and there was only 2 diffs in the “ua-list-all.txt”
file, but I confirmed that these two diffs exist in the original
crossjoin code.

They’re pretty obscure, and probably need to be cleaned up a bit anyhow:

>Sleipnir/3.3 CFNetwork/711.1.12 Darwin/14.0.0
>         browscap-php: ["Darwin","CFNetwork",""]
>   crossjoin-browscap: ["","Sleipnir","2.0"]
>
>Sleipnir/4.5.1 CFNetwork/673.3 Darwin/13.4.0 (x86_64) (iMac14%2C1)
>         browscap-php: ["Darwin","CFNetwork",""]
>   crossjoin-browscap: ["","Sleipnir","2.0"]