Improvement for manual updates
The update() method also prepares the data now. The automatic updates
are not influenced by this change, because it's the same what they did
before in three single steps. The manual updates does a bit more now -
before this change the preparation was done with the first request after
the manual update call, which is what you normally want to avoid when
they run the update manually.