Fix for #33 (not all user agents detected correctly)
The fix consists of two parts:
- First patterns that didn't contain a placeholder were intentionally
ignored, because nearly all patterns except the ones used as parents
contain a placeholder - but about 20 of them were not used as parents
and therefore filtered too...
- The other problem was the order of matches when searching the best
matching pattern - the expected order (the order in the browscap source)
isn't always correct. Now the result is first ordered by pattern lentgh
and then by position in the source.